refactor(core): fix action module imports
