+++ projects/dsp-ui/src/lib/search/fulltext-search/fulltext-search.component.ts
@@ -22,7 +22,7 @@ import {
     ProjectsResponse,
     ReadProject
 } from '@dasch-swiss/dsp-js';
-import { NotificationService } from '../../action';
+import { NotificationService } from '../../action/services/notification.service';
 import { SortingService } from '../../action/services/sorting.service';
 import { DspApiConnectionToken } from '../../core/core.module';
 import { SearchParams } from '../../viewer';
