+++ projects/dsp-ui/src/lib/viewer/views/property-view/property-toolbar/property-toolbar.component.ts
@@ -7,11 +7,10 @@ import {
     ProjectResponse,
     ReadProject,
     ReadResource,
-    ReadUser,
-    UserResponse
+    ReadUser
 } from '@dasch-swiss/dsp-js';
+import { NotificationService } from '../../../../action/services/notification.service';
 import { DspApiConnectionToken } from '../../../../core/core.module';
-import { NotificationService } from '../../../../action';
 import { UserService } from '../../../services/user.service';
 
 @Component({
